splitting `analysis` module into two: `analysis` and `cli`
