+++ settings.gradle.kts
@@ -17,3 +17,4 @@ include("jacodb-core")
 include("jacodb-analysis")
 include("jacodb-examples")
 include("jacodb-benchmarks")
+include("jacodb-cli")
